tienda_vinilos/requirements.txt
#Componentes auxiliares  a instalar

django == 3.2.9
pillow
django-filter